[⚙️ chore] 진행 중인 챌린지 컴포넌트 분리 #89
