+++ src/pages/home/index.tsx
@@ -1,23 +1,20 @@
-import Link from 'next/link';
-import Image from 'next/image';
-import styled from '@emotion/styled';
 import { useEffect, useState } from 'react';
+import styled from '@emotion/styled';
+import { GetServerSidePropsContext } from 'next';
 import { useRouter } from 'next/router';
 import { getCookie } from 'cookies-next';
-import { GetServerSidePropsContext } from 'next';
 import axios from 'axios';
 import { SLayoutWrapper } from '@/components/common/Layout';
 import Footer from '@/components/common/Footer';
-import TopBanner from '@/components/home/TopBanner';
-import ProcessingChallengeCard from '@/components/home/ProcessingChallengeCard';
-import FloatingBtn from '@/components/home/FloatingBtn';
-import HomeHeader from '@/components/home/HomeHeader';
 import SnackBar from '@/components/common/SnackBar';
+import HomeHeader from '@/components/home/HomeHeader';
+import TopBanner from '@/components/home/TopBanner';
+import FloatingBtn from '@/components/home/FloatingBtn';
+import ProcessingChallenge from '@/components/home/ProcessingChallenge';
+import RecruitingChallenge from '@/components/home/RecruitingChallenge';
 import ISnackBarState from '@/types/snackbar';
 import IRecruitingChallenge from '@/types/recruitingChallenge';
 import IMyProcessingChallenge from '@/types/myProcessingChallenge';
-import RecruitingChallenge from '@/components/home/RecruitingChallenge';
-import ScrollXBox from '@/components/common/ScrollXBox';
 import {
   getMyProcessingChallengeApi,
   getRecruitingChallengeApi,
@@ -139,29 +136,9 @@ export default function Home({
         {isLogined &&
           myProcessingChallenges &&
           myProcessingChallenges.length > 0 && (
-            <SSection>
-              <STitleLink href="/mychallenge">
-                <h2>👨‍💻 진행 중인 챌린지</h2>
-                <Image
-                  src="/icons/icon-left-arrow.svg"
-                  alt="마이 챌린지로 가기"
-                  width={24}
-                  height={24}
-                  style={{ transform: 'rotate(180deg)' }}
-                  priority
-                />
-              </STitleLink>
-              <ScrollXBox>
-                <SListScrollX>
-                  {myProcessingChallenges.map((challenge) => (
-                    <ProcessingChallengeCard
-                      key={challenge.challengeGroupId}
-                      {...challenge}
-                    />
-                  ))}
-                </SListScrollX>
-              </ScrollXBox>
-            </SSection>
+            <ProcessingChallenge
+              myProcessingChallenges={myProcessingChallenges}
+            />
           )}
         <RecruitingChallenge recruitingChallenges={recruitingChallenges} />
         {snackBarState.open && (
@@ -213,33 +190,4 @@ export const getServerSideProps = serverErrorCatch(getServerSidePropsFunction);
 const SHomeWrapper = styled(SLayoutWrapper)`
   position: relative;
   margin-bottom: 3rem;
-`;
-
-const SSection = styled.section`
-  &::after {
-    content: '';
-    display: inline-block;
-    width: 100%;
-    height: 6px;
-    margin: 1rem 0;
-    background-color: ${({ theme }) => theme.color.gray_ec};
-  }
-`;
-
-const STitleLink = styled(Link)`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: calc(100% - 2.5rem);
-  height: 60px;
-  margin: 0 1.25rem;
-  color: ${({ theme }) => theme.color.gray_3c};
-  h2 {
-    font-size: ${({ theme }) => theme.fontSize.subtitle1};
-    font-weight: ${({ theme }) => theme.fontWeight.subtitle1};
-  }
-`;
-
-const SListScrollX = styled.ul`
-  margin: 0 1.25rem;
 `;
